+++ DeathNotesPanel/DeathNotesPanel.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.ComponentModel;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using Oxide.Core.Plugins;
@@ -6,34 +7,26 @@ using UnityEngine;
 
 namespace Oxide.Plugins
 {
-    [Info("Wipe Panel", "MJSU", "0.0.5")]
-    [Description("Displays days to wipe in magic panel")]
-    internal class WipePanel : RustPlugin
+    [Info("Death Notes Panel", "MJSU", "0.0.5")]
+    [Description("Displays death notes in MagicPanel")]
+    internal class DeathNotesPanel : RustPlugin
     {
         #region Class Fields
-        [PluginReference] private readonly Plugin MagicPanel, WipeInfoApi;
+        [PluginReference] private Plugin MagicPanel;
+
         private PluginConfig _pluginConfig; //Plugin Config
-        private int _daysTillWipe;
-        
-        private string _textFormat;
+        private string _deathText;
 
         private enum UpdateEnum { All = 1, Panel = 2, Image = 3, Text = 4 }
+
+        private Timer _updateTimer;
+        private readonly Queue<string> _deathQueue = new Queue<string>();
         #endregion
 
         #region Setup & Loading
         private void Init()
         {
-            _textFormat = _pluginConfig.Panel.Text.Text;
-        }
-        
-        protected override void LoadDefaultMessages()
-        {
-            lang.RegisterMessages(new Dictionary<string, string>
-            {
-                [LangKeys.Today] = "Today",
-                [LangKeys.OneDay] = "1 Day",
-                [LangKeys.Days] = "{0} Days"
-            }, this);
+            _deathText = _pluginConfig.Panel.Text.Text;
         }
 
         protected override void LoadDefaultConfig()
@@ -58,51 +51,35 @@ namespace Oxide.Plugins
                     Enabled = config.Panel?.Image?.Enabled ?? false,
                     Color = config.Panel?.Image?.Color ?? "#FFFFFFFF",
                     Order = config.Panel?.Image?.Order ?? 0,
-                    Width = config.Panel?.Image?.Width ?? 0.4f,
+                    Width = config.Panel?.Image?.Width ?? 0.33f,
                     Url = config.Panel?.Image?.Url ?? "",
-                    Padding = config.Panel?.Image?.Padding ?? new TypePadding(0.05f, 0.0f, 0.1f, 0.1f)
+                    Padding = config.Panel?.Image?.Padding ?? new TypePadding(0.05f, 0.0f, 0.2f, 0.05f)
                 },
                 Text = new PanelText
                 {
                     Enabled = config.Panel?.Text?.Enabled ?? true,
-                    Color = config.Panel?.Text?.Color ?? "#08C717FF",
+                    Color = config.Panel?.Text?.Color ?? "#838383",
                     Order = config.Panel?.Text?.Order ?? 1,
                     Width = config.Panel?.Text?.Width ?? 1f,
                     FontSize = config.Panel?.Text?.FontSize ?? 14,
                     Padding = config.Panel?.Text?.Padding ?? new TypePadding(0.05f, 0.05f, 0.05f, 0.05f),
                     TextAnchor = config.Panel?.Text?.TextAnchor ?? TextAnchor.MiddleCenter,
-                    Text = config.Panel?.Text?.Text ?? "Wipe: {0}",
+                    Text = config.Panel?.Text?.Text ?? "{0}",
                 }
             };
             config.PanelSettings = new PanelRegistration
             {
-                BackgroundColor = config.PanelSettings?.BackgroundColor ?? "#FFF2DF08",
-                Dock = config.PanelSettings?.Dock ?? "centerupper",
-                Order = config.PanelSettings?.Order ?? 10,
-                Width = config.PanelSettings?.Width ?? 0.08f
+                BackgroundColor = config.PanelSettings?.BackgroundColor ?? "#fff2df08",
+                Dock = config.PanelSettings?.Dock ?? "undercompass",
+                Order = config.PanelSettings?.Order ?? 1,
+                Width = config.PanelSettings?.Width ?? 0.4f
             };
             return config;
         }
 
         private void OnServerInitialized()
         {
-            if (WipeInfoApi == null)
-            {
-                PrintError("Missing plugin dependency WipeInfoApi: https://github.com/dassjosh/MagicPanel");
-                return;
-            }
-            
-            OnWipeCalculated();
-        }
-        
-        private void OnWipeCalculated()
-        {
-            timer.In(1f, () =>
-            {
-                _daysTillWipe = WipeInfoApi.Call<int>("GetDaysTillWipe");
-                MagicPanel?.Call("UpdatePanel", Name, UpdateEnum.Text);
-                RegisterPanels();
-            });
+            RegisterPanels();
         }
 
         private void RegisterPanels()
@@ -112,8 +89,30 @@ namespace Oxide.Plugins
                 PrintError("Missing plugin dependency MagicPanel: https://github.com/dassjosh/MagicPanel");
                 return;
             }
-
+        
             MagicPanel?.Call("RegisterGlobalPanel", this, Name, JsonConvert.SerializeObject(_pluginConfig.PanelSettings), nameof(GetPanel));
+        }
+        #endregion
+
+        #region Death Notes Hook
+
+        private void OnDeathNotice(Dictionary<string, object> data, string message)
+        {
+            _deathQueue.Enqueue(message);
+            if (_updateTimer == null)
+            {
+                MagicPanel?.Call("UpdatePanel", Name, (int) UpdateEnum.Text);
+                _updateTimer = timer.Every(_pluginConfig.DisplayDuration, () =>
+                {
+                    MagicPanel?.Call("UpdatePanel", Name, (int) UpdateEnum.Text);
+                    
+                    if (_deathQueue.Count == 0)
+                    {
+                        _updateTimer?.Destroy();
+                        _updateTimer = null;
+                    }
+                });
+            }
         }
         #endregion
 
@@ -125,44 +124,31 @@ namespace Oxide.Plugins
             PanelText text = panel.Text;
             if (text != null)
             {
-                string message;
-                if (_daysTillWipe == 0)
+                string message = string.Empty;
+                if (_deathQueue.Count != 0)
                 {
-                    message = Lang(LangKeys.Today);
+                    message = _deathQueue.Dequeue();
                 }
-                else if (_daysTillWipe == 1)
-                {
-                    message = Lang(LangKeys.OneDay);
-                }
-                else
-                {
-                    message = Lang(LangKeys.Days, _daysTillWipe);
-                }
-
-                text.Text = string.Format(_textFormat, message);
+                
+                text.Text = string.Format(_deathText, message);
             }
 
             return JsonConvert.SerializeObject(panel);
         }
-        
-        private string Lang(string key, params object[] args) => string.Format(lang.GetMessage(key, this), args);
         #endregion
 
         #region Classes
         private class PluginConfig
         {
+            [DefaultValue(5f)]
+            [JsonProperty(PropertyName = "Display Duration (Seconds)")]
+            public float DisplayDuration { get; set; }
+
             [JsonProperty(PropertyName = "Panel Settings")]
             public PanelRegistration PanelSettings { get; set; }
 
             [JsonProperty(PropertyName = "Panel Layout")]
             public Panel Panel { get; set; }
-        }
-        
-        private class LangKeys
-        {
-            public const string Today = "Today";
-            public const string OneDay = "OneDay";
-            public const string Days = "Days";
         }
 
         private class PanelRegistration