+++ PanelRequiredPlugins/WipeInfoApi.cs
@@ -1,52 +1,44 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.Linq;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using Oxide.Core;
-using Oxide.Core.Libraries.Covalence;
-using Oxide.Core.Plugins;
-using UnityEngine;
+using Rust;
 
 namespace Oxide.Plugins
 {
-    [Info("Wipe Info", "MJSU", "0.0.1")]
-    [Description("Tells players when the last and next wipe is")]
-    internal class WipeInfo : RustPlugin
+    [Info("Wipe Info Api", "MJSU", "0.0.1")]
+    [Description("Api for when the server is wiping")]
+    internal class WipeInfoApi : RustPlugin
     {
         #region Class Fields
-        [PluginReference] private Plugin DiscordCore;
-
+        private StoredData _storedData; //Plugin Data
         private PluginConfig _pluginConfig;
 
         private DateTime _nextWipe;
         private int _daysTillNextWipe;
+        private bool _isForcedWipeDay;
+        private bool _newSaveVersion;
         #endregion
 
         #region Setup & Loading
+
         private void Init()
         {
-            for (int i = 0; i < _pluginConfig.AutoChatTrigger.Count; i++)
+            _storedData = Interface.Oxide.DataFileSystem.ReadObject<StoredData>(Name);
+
+            if (_storedData.SaveVersion != Protocol.save)
             {
-                _pluginConfig.AutoChatTrigger[i] = _pluginConfig.AutoChatTrigger[i].ToLower();
+                _storedData.SaveVersion = Protocol.save;
+                _newSaveVersion = true;
+                Interface.Call("OnSaveVersionChanged");
+                NextTick(SaveData);
             }
             
-            cmd.AddChatCommand(_pluginConfig.GameChatCommand, this, WipeChatCommand);
+            CalculateWipe();
         }
-        
-        protected override void LoadDefaultMessages()
-        {
-            lang.RegisterMessages(new Dictionary<string, string>
-            {
-                [LangKeys.Today] = "Today",
-                [LangKeys.InDays] = "in {0} days",
-                [LangKeys.WipeText] = "The next wipe will occur {0} on {1:dddd MM/dd/yyyy}. " +
-                                      "Last wipe was {2:MM/dd/yyyy}.",
-                [LangKeys.Chat] = $"<color=#BEBEBE>[<color=#de8732>{Title}</color>] {{0}}</color>",
-            }, this);
-        }
-        
+
         protected override void LoadDefaultConfig()
         {
             PrintWarning("Loading Default Config");
@@ -79,14 +71,6 @@ namespace Oxide.Plugins
                     Occurrence = 1
                 }
             };
-
-            config.AutoChatTrigger = config.AutoChatTrigger ?? new List<string>
-            {
-                "wipe",
-                "wipe?",
-                "wiped",
-                "wiped?"
-            };
             return config;
         }
 
@@ -98,43 +82,19 @@ namespace Oxide.Plugins
         private void OnServerInitialized()
         {
             CalculateWipe();
-            OnDiscordCoreReady();
-
-            if (!_pluginConfig.EnableAutoChatResponse)
-            {
-                Unsubscribe(nameof(OnPlayerChat));
-            }
-        }
-
-        private void OnDiscordCoreReady()
-        {
-            timer.In(1f, () =>
-            {
-                DiscordCore?.Call("RegisterCommand", _pluginConfig.DiscordChatCommand, this, new Func<IPlayer, string, string, string[], object>(HandleWipe), "To show days till wipe", null, true); 
-            });
-        }
-        
-        private void OnPlayerInit(BasePlayer player)
-        {
-            if (player.HasPlayerFlag(BasePlayer.PlayerFlags.ReceivingSnapshot))
-            {
-                timer.Once(5f, () => OnPlayerInit(player));
-                return;
-            }
-
-            timer.Once(10f, () => { WipeChatCommand(player, _pluginConfig.GameChatCommand, new string[0]); });
-        }
-
-        private void Unload()
-        {
-            DiscordCore?.Call("UnregisterCommand", _pluginConfig.DiscordChatCommand, this);
         }
         #endregion
 
         #region Wipe Calculations
         private void CalculateWipe()
         {
-            DateTime wipeDate = SaveRestore.SaveCreatedTime.Date;
+            if (SaveRestore.SaveCreatedTime != DateTime.MinValue && _storedData.PreviousWipe != SaveRestore.SaveCreatedTime.Date)
+            {
+                _storedData.PreviousWipe = SaveRestore.SaveCreatedTime.Date;
+                NextTick(SaveData);
+            }
+            
+            DateTime wipeDate = _storedData.PreviousWipe;
 
             DateTime lastWipe = GetWipe(wipeDate);
             DateTime nextWipe = GetWipe(wipeDate.AddMonths(1));
@@ -143,6 +103,11 @@ namespace Oxide.Plugins
             {
                 nextWipe = lastWipe;
                 lastWipe = GetWipe(wipeDate.AddMonths(-1));
+            }
+
+            if (nextWipe.Date == DateTime.Now.Date && !_newSaveVersion)
+            {
+                _isForcedWipeDay = true;
             }
 
             List<WipeSchedule> schedule = null;
@@ -214,82 +179,51 @@ namespace Oxide.Plugins
 
         #region API Hooks
 
-        private int DaysTillWipe()
+        private int GetDaysTillWipe()
         {
             return _daysTillNextWipe;
         }
 
-        private int DaysBetweenWipes()
+        private DateTime GetNextWipe()
+        {
+            return _nextWipe;
+        }
+
+        private int GetDaysBetweenWipe()
         {
             return (_nextWipe - SaveRestore.SaveCreatedTime).Days;
         }
 
-        private string GetWipeText(string id = null)
+        private bool IsForcedWipeDay()
         {
-            string nextWipe = _daysTillNextWipe == 0 ? Lang(LangKeys.Today, id) : Lang(LangKeys.InDays, id, _daysTillNextWipe);
-
-            return Lang(LangKeys.WipeText, id, nextWipe, _nextWipe, SaveRestore.SaveCreatedTime.ToLocalTime());
-        }
-        #endregion
-
-        #region Chat Command
-        private void WipeChatCommand(BasePlayer player, string command, string[] args)
-        {
-            Chat(player, GetWipeText(player.UserIDString));
+            return _isForcedWipeDay;
         }
 
-        private void OnPlayerChat(ConsoleSystem.Arg arg)
+        private bool IsNewSaveVersion()
         {
-            BasePlayer player = arg.Player();
-            if (_pluginConfig.AutoChatTrigger.Contains(arg.FullString.ToLower()))
-            {
-                WipeChatCommand(player, null, null);
-            }
-        }
-        #endregion
-
-        #region Discord Command
-        private object HandleWipe(IPlayer player, string channelId, string cmd, string[] args)
-        {
-            SendMessage(channelId, GetWipeText(player.Id));
-            return null;
+            return _newSaveVersion;
         }
         #endregion
 
         #region Helper Methods
-        private void SendMessage(string channelId, string message)
-        {
-            DiscordCore?.Call("SendMessageToChannel", channelId, $"{Title}: {message}");
-        }
-
-        private string Lang(string key, string id, params object[] args) => string.Format(lang.GetMessage(key, this, id), args);
-        private void Chat(BasePlayer player, string format) => PrintToChat(player, Lang(LangKeys.Chat, player.UserIDString, format));
+        private void SaveData() => Interface.Oxide.DataFileSystem.WriteObject(Name, _storedData);
         #endregion
 
         #region Classes
 
         private class PluginConfig
         {
-            [DefaultValue(true)]
-            [JsonProperty("Enable Auto Chat")]
-            public bool EnableAutoChatResponse { get; set; }
-            
-            [JsonProperty("Auto Chat Triggers")]
-            public List<string> AutoChatTrigger { get; set; }
-            
-            [DefaultValue("wipe")]
-            [JsonProperty("Game Chat Command")]
-            public string GameChatCommand { get; set; }
-            
-            [DefaultValue("wipe")]
-            [JsonProperty("Discord Chat Command")]
-            public string DiscordChatCommand { get; set; }
-            
             [JsonProperty("4 week schedule")]
             public List<WipeSchedule> ScheduleWeek4 { get; set; }
             
             [JsonProperty("5 week schedule")]
             public List<WipeSchedule> ScheduleWeek5 { get; set; }
+        }
+
+        private class StoredData
+        {
+            public int SaveVersion { get; set; }
+            public DateTime PreviousWipe { get; set; } = DateTime.MinValue;
         }
 
         private class WipeSchedule
@@ -300,14 +234,6 @@ namespace Oxide.Plugins
             
             [JsonProperty("Day of the week occurrence (1-5)")]
             public int Occurrence { get; set; }
-        }
-        
-        private class LangKeys
-        {
-            public const string Today = "Today";
-            public const string InDays = "InDays";
-            public const string WipeText = "WipeText";
-            public const string Chat = "Chat";
         }
         #endregion
     }